requirements.txt
openai>=1.88
langchain-community>=0.3.23
langchain-mcp-adapters>=0.0.9
langchain-openai>=0.2.14
mcp>=1.6.0
mem0ai>=0.1.94
python-dotenv>=1.1.0
supabase>=2.0.0
unstructured>=0.17.2
psycopg2-binary>=2.9.9
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
firebase-admin>=6.4.0
pytz>=2024.1
vecs>=0.1.0
crewai==0.141.0
requests>=2.31.0 